mcp-servers/api/requirements.txt
mcp
sse-starlette
starlette
google-adk==1.8.0